RaspberryPi : add early PiDmaLib
This is a modified fork of NonCoherentDmaLib, that uses
AllocateMaxAddress <= 1GB allocations instead of generic
page/pool allocs. It also switches the cached buffer bouncing
to use page allocs instead of the awkward pool allocs, so
that's a bonus.

Note: this does *not* bounce requests to map host addresses
> 1GB. None of the code for the Pi do that - they allocates
DMA'able buffers instead using the DmaLib. The only drivers
using DmaLib today are:
- DWC2 driver
- Display
- Mailbox

If/when SDMMC drivers decide to support DMA to arbitrary
buffers (why not), this should be trivial to add. Really,
this entire driver could be replaced by extending the EmbeddedPkg
NonCoherentDmaLib to support a max address PCD.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>