Silicon/Broadcom/Bcm283x: Add PiDmaLib
This is a modified fork of NonCoherentDmaLib, that uses
AllocateMaxAddress <= 1GB allocations instead of generic
page/pool allocs. It also switches the cached buffer bouncing
to use page allocations instead of pool allocations.

Note: This does *not* bounce requests to map host addresses
above 1GB. None of the code for the Pi does that as it allocates
DMA'able buffers instead using the DmaLib.
The only drivers using DmaLib today are:
- DWC2 driver
- Display
- Mailbox

If/when SDMMC drivers decide to support DMA to arbitrary buffers
this should be trivial to add. As a matter of fact, this entire
driver could be replaced by extending the EmbeddedPkg
NonCoherentDmaLib to support a max address PCD.

Signed-off-by: Samer El-Haj-Mahmoud <samer.el-haj-mahmoud@arm.com>>