Platform/RPi4: switch to 5.1 version MADT
Older versions of the Linux kernel are finicky when it comes to
checking the size of the GICC entries in the MADT table: depending
on the FADT version (either 5, or 6 or later), the size must be
exactly 76 or 80 bytes, respectively. Using the ACPI 6.0 or 6.1
GICC macros results in 80 byte entries, which triggers a mismatch
error since the FADT we are exposing is 5.1.

Since GICv2 doesn't care about any of the fields that were added
after ACPI 5.1, just switch our MADT table back to it and make
things match again.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>