Silicon/Bcm283x: Add FIFO mode for RNG
The Bcm283x Random Number Generator does not work in regular mode
on the Bcm2711 powered Raspberry Pi 4. It does however work when
using FIFO mode. So we add this new mode, which is governed by the
use of the new PcdBcm283xRngUseFifo boolean PCD.

Note that just as a Pi 4 doesn't seem to support regular mode, a
Pi 3 doesn't seem to support FIFO mode, which is why we can't
switch to simply using FIFO mode for both platforms.

We also fix the register to which RNG_WARMUP_COUNT is written,
which was incorrect.

Signed-off-by: Pete Batard <pete@akeo.ie>