Platform/RPi4: switch to the PL011 for DEBUG output and console
The PL011 is a better choice for the serial console on the RPi4,
given that its baud clock is not derived from the CPU clock (which
may change under our feet unless we keep it fixed at a low rate),
and give the fact that the SBSA/SBBR specs that describe ARM specific
architectural requirements for ACPI only permit PL011 based UARTs
to begin with.

So switch the build to PL011 by default for DEBUG output as well as
the UEFI console and the SPCR console. Also, fix the interrupt line
in the DSDT and set the _CID to ARMH0011 so it gets picked up by a
SBSA/SBBR compliant OS.

Note that this requires an overlay to be enabled in config.txt that
pinmuxes the PL011 TX/RX lines to the UART pins on the connector
block.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>