Platform/RPi4: Build the PCIe and xHCI drivers into the firmware
This uses the recently introduced NonCoherentIoMmuDxe for PCIe access
and should enable USB device usage in the UEFI environment.

As mentioned in https://lkml.org/lkml/2019/9/9/170, imposing a 3 GB
DMA limit might be necessary for the 4 GB models so we follow suit by
setting PcdDmaDeviceLimit to 3 GB - 1 in NonCoherentIoMmuDxe.

Note that this patch does not provide xHCI ACPI support because the
required Xhci.asl table will be provided in a later commit.

Signed-off-by: Pete Batard <pete@akeo.ie>