RPi4: describe DMA constraints for all the legacy devices behind GPU
In the process:

- Comment out GPU and PWM devs. GPU needs to be
validate if its even compatible enough (and interrupt is wrong
anyway. PWM is just sketchy as _CRS encodes CPU and bus
addresses which isn't going to work (and this should be done
via ACPI0004 or similar)

- Since references (e.g SB.GPIO0) changed, use this as an
opportunity to replace MsftFunctionConfig/VendorLong with
PinFunction (the proper ACPI standard way) See old
https://github.com/andreiw/RaspberryPiPkg/pull/134/commits
for more info.

Oh and completely untested. This commit is for the NetBSD folks,
seeing that they are leading the world on implemneting DMA
constraints/translation support in ACPI.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>