Platform/RPi4: Use Silicon constants in ACPI headers
With the new constants added, remove some of the hardcoded values from
the ACPI tables.

Note that because the ASL compiler is very limited in terms of macro
processing, we can not use any arithmetic constants (e.g BASE + OFFSET)
as parameters to MEMORY32FIXED (), and instead must alter the base
address using CreateDwordField () after MEMORY32FIXED () has been
invoked.

To achieve that, we create a MEMORY32SETBASE () macro and move the
MEMORY32FIXED () calls outside of Method (_CRS...).

Signed-off-by: Pete Batard <pete@akeo.ie>