Platform/RPi3: Use proper aslc for FADT, GTDT and MADT tables generation
The FADT and GTDT tables are virtually identical to the ones we use for
the Pi 4, except for the RPI_SYSTEM_TIMER_BASE_ADDRESS in Gtdt.aslc.

However, 2 changes are required for these tables to work with Windows:
1. The ACPI OEM ID for the FADT table MUST be "BC2836" (else Windows 10
   fails to boot) so we change the global value in AcpiTables.h.
2. The values for PcdArmArchTimerVirtIntrNum and PcdArmArchTimerHypIntrNum
   seem to have been swapped in our existing .dsc so we swap them back
   (Without this swap, Windows 10 simply refuses to boot).

The MADT table is deduced from the original Microsoft blob. The one thing
of notice is that no GIC Distributor section should exist in that table,
otherwise Windows 10 will fail to boot.

Obviously, all of the changes applied by this commit were validated on a
Raspberry Pi 3 running Windows 10.

Signed-off-by: Pete Batard <pete@akeo.ie>