Platform/RPi4: Add RPI_MODEL constant and replace PL011_ENABLE
Since there are multiple Raspberry Pi Models and a lot of the code we use
can be factorized, it is useful to have an RPI_MODEL build time constant,
set to the platform model number, that can be referenced in the source.

Make use this new constant to replace the PL011_ENABLE feature check.

Signed-off-by: Pete Batard <pete@akeo.ie>