Platform/RPi/DualSerialPortLib: Fix miniUART serial divisor computation
Fix for https://github.com/raspberrypi/firmware/issues/1376.

For the Pi 3, to properly configure miniUART, we need the core clock,
which can be vary between VideoCore firmare release or config.txt options.

Unfortunately, it's painful to get - it's only available via the mailbox
interface. Additionally, SerialLib is a very limited environment, even
when linked in a DXE-mode component, because as part of a DebugLib
implementation it ends being the base prerequisite of everything.
That means a "normal" mailbox implementation like the one from
RpiFirmwareDxe (with dependencies on DmaLib) is out of the question.
Using a basic implementation such as the one in PlatformLib doesn't work
either because it operates in both environments with MMU on (DXE phase)
and MMU off (SEC/PrePi).

Ideally, we read the value via mbox exactly once at boot (PlatformLib),
and then somehow stash it. A GUID Hob sounds appropriate, yet when
SerialPortLib operates in DXE components, we can't use the HobLib to
*locate* the Hob list itself (remember, SerialPortLib initializes
before any of HobLib's dependencies, like UeflLib...).

FORTUNATELY, there is a way out - we can use PrePeiGetHobList to cut
through to the HOB list pointer stashed by PrePi without dealing with
any of the libraries meant for DXE.

Once we have the Hob list pointer, we can use the regular DXE HobLib
routines safely (so long as they are the ones taking a HobList pointer
as a parameter). So we can link DualSerialPortLib with whatever HobLib
that is appropriate for the environment - just need to also link in
PrePiHobListPointerLib.

gSerialLibCoreClockFreq is always externally set when DualSerialPortLib
run as part of PrePi. So - if gSerialLibCoreClockFreq is not set, then
we must be in the DXE phase and can fetch the Hob list via
PrePeiGetHobList. The GUID Hob is registered as part of a Pi-customized
PlatformPeiLib.

For the Pi 4, the situation is a bit different, as the system clock used
to compute the baud rate is not derived directly from the core clock, but
instead from a fixed clock (current 1 GHz) to which a 12.12 fixed point
divisor is applied. But even though we don't use it, we still initialize
gSerialLibCoreClockFreq a.k.a. the VPU core frequency, as this can be
useful to have.

As a result of the above, the PcdSerialClockRate values in the .dsc are
also updated, so that it can be used as a fallback to the default core
frequency in case of the Pi 3, and as the fixed frequency to which the
fixed point divisor is applied in case of the Pi 4.

Co-authored-by: Pete Batard <pete@akeo.ie>
Co-authored-by: Andrei Warkentin <andrey.warkentin@gmail.com>
Signed-off-by: Pete Batard <pete@akeo.ie>